requirements.txt
fastapi
uvicorn
pypdf
chromadb
google-genai
python-multipart
pydantic
streamlit
